Remove TestInput from TVRenamerTest and make it stand alone class
 Changes to be committed:
	modified:   src/test/org/tvrenamer/TVRenamerTest.java
	new file:   src/test/org/tvrenamer/TestInput.java
