+++ src/test/org/tvrenamer/TVRenamerTest.java
@@ -148,24 +148,4 @@ public class TVRenamerTest {
             fail(e.getMessage());
         }
     }
-
-    private static class TestInput {
-        public final String input;
-        public final String show;
-        public final String season;
-        public final String episode;
-
-        public final String episodeTitle;
-        public final String episodeResolution;
-
-        public TestInput(String input, String show, String season, String episode, String episodeTitle, String episodeResolution) {
-            this.input = input;
-            this.show = show.toLowerCase();
-            this.season = season;
-            this.episode = episode;
-            this.episodeTitle = episodeTitle;
-            this.episodeResolution = episodeResolution;
-        }
-    }
-
 }
